pyproject.toml
[tool.poetry]
name = "kultunaut-lib"
version = "0.1.0"
description = ""
authors = ["nbnotabene <niels@bjerre.net>"]
readme = "README.md"
packages = [{include = "kultunaut"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
mariadb = "^1.1.11"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
